Performace improvement to `watch` command.
fix `watch` from idling at 12% CPU usage. I was able to
fix the bottleneck with a cProfile. This is before the fix https://svgur.com/s/e71
and this is after https://svgur.com/s/e8b.

The issue was that `sleep` was inside the if branch which makes
the `squirrel` test `if event_handler.files:` over and over without
sleeping.